+++ csi.c
@@ -569,6 +569,12 @@ decset_decrst(struct terminal *term, unsigned param, bool enable)
         term->report_theme_changes = enable;
         break;
 
+    case 2033:
+        term->visibility_reports = enable;
+        if (enable)
+            term_send_visibility_report(term);
+        break;
+
     case 2048:
         if (enable)
             term_enable_size_notifications(term);
@@ -664,6 +670,7 @@ decrqm(const struct terminal *term, unsigned param)
         ? DECRPM_PERMANENTLY_RESET
         : decrpm(term->grapheme_shaping);
     case 2031: return decrpm(term->report_theme_changes);
+    case 2033: return decrpm(term->visibility_reports);
     case 2048: return decrpm(term->size_notifications);
     case 8452: return decrpm(term->sixel.cursor_right_of_graphics);
     case 737769: return decrpm(term_ime_is_enabled(term));
@@ -710,6 +717,7 @@ xtsave(struct terminal *term, unsigned param)
     case 2026: term->xtsave.app_sync_updates = term->render.app_sync_updates.enabled; break;
     case 2027: term->xtsave.grapheme_shaping = term->grapheme_shaping; break;
     case 2031: term->xtsave.report_theme_changes = term->report_theme_changes; break;
+    case 2033: term->xtsave.visibility_reports = term->visibility_reports; break;
     case 2048: term->xtsave.size_notifications = term->size_notifications; break;
     case 8452: term->xtsave.sixel_cursor_right_of_graphics = term->sixel.cursor_right_of_graphics; break;
     case 737769: term->xtsave.ime = term_ime_is_enabled(term); break;
@@ -755,6 +763,7 @@ xtrestore(struct terminal *term, unsigned param)
     case 2026: enable = term->xtsave.app_sync_updates; break;
     case 2027: enable = term->xtsave.grapheme_shaping; break;
     case 2031: enable = term->xtsave.report_theme_changes; break;
+    case 2033: enable = term->xtsave.visibility_reports; break;
     case 2048: enable = term->xtsave.size_notifications; break;
     case 8452: enable = term->xtsave.sixel_cursor_right_of_graphics; break;
     case 737769: enable = term->xtsave.ime; break;
@@ -1590,6 +1599,10 @@ csi_dispatch(struct terminal *term, uint8_t final)
                 term_to_slave(term, reply, chars);
                 break;
             }
+
+            case 998:
+                term_send_visibility_report(term);
+                break;
             }
             break;
         }